Added strange shattered + star
Also reorganized the planet.assets into a single file since they don't
need to be seperate and it becomes much harder to read if they are.

Set the groundwork for adding the strangeconvert planet class but
I'll leave its visuals more upto spartan. and if any of you would like
me to change the particles or visuals of either the strange star or
shattered planet let me know on discord and I'll consider making Changes